+++ scanner/src/main/java/ca/on/oicr/gsi/runscanner/scanner/processor/V2PacBioProcessor.java
@@ -36,8 +36,8 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.w3c.dom.Document;
 
-/** Scan PacBio Revio runs from a directory. */
-public class RevioPacBioProcessor extends RunProcessor {
+/** Scan PacBio Revio and Vega runs from a directory. */
+public class V2PacBioProcessor extends RunProcessor {
 
   /** Extract data from an XML metadata file and put it in the DTO. */
   interface ProcessMetadata {
@@ -50,27 +50,26 @@ public class RevioPacBioProcessor extends RunProcessor {
       Pattern.compile("[0-9]_[A-Z][0-9]{2}").asPredicate();
 
   private static final Predicate<String> TRANSFER_TEST =
-      Pattern.compile("Transfer_Test_[0-9]{6}_[0-9]{6}.txt").asPredicate();
+      Pattern.compile("Transfer_Test_.*\\.txt").asPredicate();
 
-  private static final Predicate<String> TRANSFER_DONE =
-      Pattern.compile("[a-z][0-9]{5}_[0-9]{6}_[0-9]{6}_s[0-9].transferdone").asPredicate();
+  private static final String TRANSFER_DONE_SUFFIX = ".transferdone";
 
-  private static final Predicate<String> PB_REPORT_LOG =
-      Pattern.compile("[a-z][0-9]{5}_[0-9]{6}_[0-9]{6}_s[0-9].pbreports.log").asPredicate();
+  private static final String PB_REPORT_FILE_SUFFIX = ".pbreports.log";
 
-  private static final Logger log = LoggerFactory.getLogger(RevioPacBioProcessor.class);
+  private static final Logger log = LoggerFactory.getLogger(V2PacBioProcessor.class);
 
   private static final Pattern RUN_DIRECTORY = Pattern.compile("^.+_\\d+$");
 
   // Run information extracted from metadata XML file
-  private static final RevioPacBioProcessor.ProcessMetadata[] REVIO_METADATA_PROCESSORS =
-      new RevioPacBioProcessor.ProcessMetadata[] {
+  private static final V2PacBioProcessor.ProcessMetadata[] REVIO_METADATA_PROCESSORS =
+      new V2PacBioProcessor.ProcessMetadata[] {
         processString("//RunDetails/TimeStampedName", PacBioNotificationDto::setRunAlias),
         processString(
             "//CollectionMetadata/@InstrumentName", PacBioNotificationDto::setSequencerName),
         processDate("//Run/@WhenStarted", PacBioNotificationDto::setStartDate),
         processString(
-            "//VersionInfo[@Name='smrtlink']/@Version", PacBioNotificationDto::setSoftware),
+            "//VersionInfo[@Name='smrtlink']/@Version",
+            (dto, string) -> dto.setSoftware("SMRT Link " + string)),
         processSampleInformation()
       };
 
@@ -81,7 +80,7 @@ public class RevioPacBioProcessor extends RunProcessor {
    * @param expression the XPath expression yielding the date
    * @param setter the writer for the date
    */
-  private static RevioPacBioProcessor.ProcessMetadata processDate(
+  private static V2PacBioProcessor.ProcessMetadata processDate(
       String expression, BiConsumer<PacBioNotificationDto, Instant> setter) {
     XPathExpression expr = RunProcessor.compileXPath(expression)[0];
     return (document, dto, timeZone) -> {
@@ -95,7 +94,7 @@ public class RevioPacBioProcessor extends RunProcessor {
   }
 
   // Revio Samples
-  private static RevioPacBioProcessor.ProcessMetadata processSampleInformation() {
+  private static V2PacBioProcessor.ProcessMetadata processSampleInformation() {
     XPathExpression[] expr =
         RunProcessor.compileXPath(
             "//ResultsFolder",
@@ -134,7 +133,7 @@ public class RevioPacBioProcessor extends RunProcessor {
    * @param setter writer for the string
    * @return
    */
-  private static RevioPacBioProcessor.ProcessMetadata processString(
+  private static V2PacBioProcessor.ProcessMetadata processString(
       String expression, BiConsumer<PacBioNotificationDto, String> setter) {
     XPathExpression expr = RunProcessor.compileXPath(expression)[0];
     return (document, dto, timeZone) -> {
@@ -145,12 +144,12 @@ public class RevioPacBioProcessor extends RunProcessor {
     };
   }
 
-  public RevioPacBioProcessor(Builder builder) {
+  public V2PacBioProcessor(Builder builder) {
     super(builder);
   }
 
   public static RunProcessor create(Builder builder, ObjectNode parameters) {
-    return new RevioPacBioProcessor(builder);
+    return new V2PacBioProcessor(builder);
   }
 
   @Override
@@ -182,19 +181,22 @@ public class RevioPacBioProcessor extends RunProcessor {
 
     // Grab the .metadata.xml and begin processing
     streamSmrtCellSubdirectories(runDirectory, "metadata")
-        .flatMap(
+        .map(
             metadataDirectory ->
-                Stream.of(metadataDirectory.listFiles())
-                    .filter(
-                        file ->
-                            file.getName()
-                                .matches("^([A-Za-z0-9]+(_[A-Za-z0-9]+){3})\\.metadata\\.xml$")))
-        .map(RunProcessor::parseXml)
+                Stream.of(
+                        metadataDirectory.listFiles(
+                            (dir, name) -> {
+                              return name.endsWith(".metadata.xml") && !name.contains("preview");
+                            }))
+                    .findAny())
         .filter(Optional::isPresent)
+        .map(Optional::get)
+        .map(RunProcessor::parseXml)
         .forEach(metadata -> processMetadata(metadata.get(), dto, tz));
 
     // When a run first starts, we can only get the run alias from the directory.
-    // We use this check to ensure the same run doesn't appear under a different name when
+    // We use this check to ensure the same run doesn't appear under a different
+    // name when
     // the metadata files are written out and available
     if (dto.getRunAlias() != null) {
       // Not valid, if it doesn't match run directory name
@@ -207,7 +209,8 @@ public class RevioPacBioProcessor extends RunProcessor {
       dto.setRunAlias(runDirectory.getName());
     }
 
-    // We don't have a start date from metadata, fallback to Transfer_Test file creation time
+    // We don't have a start date from metadata, fallback to Transfer_Test file
+    // creation time
     if (dto.getStartDate() == null) {
       dto.setStartDate(startTimeFromTransferTest(runDirectory));
     }
@@ -224,8 +227,8 @@ public class RevioPacBioProcessor extends RunProcessor {
                   statisticsDirectory ->
                       Stream.of(
                           statisticsDirectory.listFiles(
-                              file -> PB_REPORT_LOG.test(file.getName()))))
-              .map(RevioPacBioProcessor::getLogCompletionTime)
+                              (dir, file) -> file.endsWith(PB_REPORT_FILE_SUFFIX))))
+              .map(V2PacBioProcessor::getLogCompletionTime)
               .max(Comparator.naturalOrder());
       // Set completion time based on pbreports.log file
       latestCompletionTime.ifPresent(dto::setCompletionDate);
@@ -250,7 +253,7 @@ public class RevioPacBioProcessor extends RunProcessor {
    * @param dto the DTO to update
    */
   private void processMetadata(Document metadata, PacBioNotificationDto dto, TimeZone timeZone) {
-    for (RevioPacBioProcessor.ProcessMetadata processor : REVIO_METADATA_PROCESSORS) {
+    for (V2PacBioProcessor.ProcessMetadata processor : REVIO_METADATA_PROCESSORS) {
       try {
         processor.accept(metadata, dto, timeZone);
       } catch (XPathException e) {
@@ -304,8 +307,7 @@ public class RevioPacBioProcessor extends RunProcessor {
    * @return completion time
    */
   private static Instant getLogCompletionTime(File file) {
-    try {
-      Scanner myReader = new Scanner(file);
+    try (Scanner myReader = new Scanner(file)) {
       Pattern pattern =
           Pattern.compile("^\\[INFO] (.*) \\[.*] exiting with return code \\d+" + " .*");
       while (myReader.hasNextLine()) {
@@ -331,23 +333,15 @@ public class RevioPacBioProcessor extends RunProcessor {
    * @return true or false
    */
   private boolean isRunComplete(File runDirectory, int smrtCellCount) throws IOException {
-    try (Stream<Path> testFileStream = Files.walk(runDirectory.toPath());
-        Stream<Path> testDoneStream = Files.walk(runDirectory.toPath())) {
-      int transferTestFiles =
-          (int)
-              testFileStream
-                  .filter(Files::isRegularFile)
-                  .filter(file -> TRANSFER_TEST.test(String.valueOf(file.getFileName())))
-                  .count();
-
+    try (Stream<Path> testDoneStream = Files.walk(runDirectory.toPath())) {
       int transferDoneFiles =
           (int)
               testDoneStream
                   .filter(Files::isRegularFile)
-                  .filter(file -> TRANSFER_DONE.test(String.valueOf(file.getFileName())))
+                  .filter(file -> file.getFileName().toString().endsWith(TRANSFER_DONE_SUFFIX))
                   .count();
 
-      return transferDoneFiles == smrtCellCount && transferTestFiles == smrtCellCount;
+      return transferDoneFiles == smrtCellCount;
     }
   }
 
@@ -358,21 +352,22 @@ public class RevioPacBioProcessor extends RunProcessor {
    * @return earliest file creation time or null if no Transfer_Test files are found
    */
   private Instant startTimeFromTransferTest(File runDirectory) throws IOException {
-    Stream<Path> testFileStream = Files.walk(runDirectory.toPath());
-    List<Path> transferTestFiles =
-        testFileStream
-            .filter(Files::isRegularFile)
-            .filter(file -> TRANSFER_TEST.test(String.valueOf(file.getFileName())))
-            .toList();
+    try (Stream<Path> testFileStream = Files.walk(runDirectory.toPath())) {
+      List<Path> transferTestFiles =
+          testFileStream
+              .filter(Files::isRegularFile)
+              .filter(file -> TRANSFER_TEST.test(String.valueOf(file.getFileName())))
+              .toList();
 
-    Instant minInstant = null;
-    for (Path filepath : transferTestFiles) {
-      Instant creationTime = getFileCreationTime(filepath.toFile());
-      if (minInstant == null || creationTime.isBefore(minInstant)) {
-        minInstant = creationTime;
+      Instant minInstant = null;
+      for (Path filepath : transferTestFiles) {
+        Instant creationTime = getFileCreationTime(filepath.toFile());
+        if (minInstant == null || creationTime.isBefore(minInstant)) {
+          minInstant = creationTime;
+        }
       }
+      return minInstant;
     }
-    return minInstant;
   }
 
   /**
@@ -382,20 +377,21 @@ public class RevioPacBioProcessor extends RunProcessor {
    * @return latest file creation time or null if no Transferdone files are found
    */
   private Instant completionTimeFromTransferDone(File runDirectory) throws IOException {
-    Stream<Path> testFileStream = Files.walk(runDirectory.toPath());
-    List<Path> transferDoneFiles =
-        testFileStream
-            .filter(Files::isRegularFile)
-            .filter(file -> TRANSFER_DONE.test(String.valueOf(file.getFileName())))
-            .toList();
+    try (Stream<Path> testFileStream = Files.walk(runDirectory.toPath())) {
+      List<Path> transferDoneFiles =
+          testFileStream
+              .filter(Files::isRegularFile)
+              .filter(file -> file.getFileName().toString().endsWith(TRANSFER_DONE_SUFFIX))
+              .toList();
 
-    Instant maxInstant = null;
-    for (Path filepath : transferDoneFiles) {
-      Instant creationTime = getFileCreationTime(filepath.toFile());
-      if (maxInstant == null || creationTime.isAfter(maxInstant)) {
-        maxInstant = creationTime;
+      Instant maxInstant = null;
+      for (Path filepath : transferDoneFiles) {
+        Instant creationTime = getFileCreationTime(filepath.toFile());
+        if (maxInstant == null || creationTime.isAfter(maxInstant)) {
+          maxInstant = creationTime;
+        }
       }
+      return maxInstant;
     }
-    return maxInstant;
   }
 }